+++ CHANGELOG.md
@@ -59,7 +59,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### API Breaking Changes
 
 * (x/staking/types) [\#7447](https://github.com/cosmos/cosmos-sdk/issues/7447)  Remove bech32 PubKey support
-  * `ValidatorI` inteface update. `GetConsPubKey` and `GetConsAddr` methods return error
+  * `ValidatorI` inteface update. `GetConsPubKey`, `GetConsAddr` methods return error
+  * `Validator` update. Methods changed in `ValidatorI` (as described above) and `ToTmValidator` return error.
   * `Validator.ConsensusPubkey` type changed from `string` to `codectypes.Any`
   * `MsgCreateValidator.Pubkey` type changed from `string` to `codectypes.Any`